ASoC: aess: update abe_typ.h for size optimization
Update structure fields size in order to decrease FW header
size

+++ sound/soc/omap/aess/abe_typ.h
@@ -428,8 +428,8 @@ struct omap_aess_dma_offset {
  */
 
 struct omap_aess_task {
-	u32 frame;
-	u32 slot;
+	u8 frame;
+	u8 slot;
 	u16 task;
 };
